Add web for adding employee

+++ HappyFarmProject/HappyFarmProjectAPI/App_Code/Helper.cs
@@ -66,7 +66,7 @@ namespace HappyFarmProjectAPI
         {
             try
             {
-                int.Parse(phoneNumber);
+                long.Parse(phoneNumber);
                 return true;
             }
             catch (Exception ex)
